Compiling 64bit SPARCV9 revealed a type mismatch (size_t vs. uint32_t), this is the fix.
git-svn-id: svn://svn.code.sf.net/p/bogofilter/code/trunk@1853 046ba1d7-8659-4f4f-8137-08d775d845e8

+++ bogofilter/src/bogoutil.c
@@ -448,8 +448,8 @@ static int display_words(const char *path, int argc, char **argv, bool prob)
 
 static double compute_robx(void *dbh_spam, void *dbh_good)
 {
-    size_t good_cnt = 0;
-    size_t spam_cnt = 0;
+    uint32_t good_cnt = 0;
+    uint32_t spam_cnt = 0;
     double sum = 0.0;
     double robx;
